checkpoint - Clean files up
