+++ src/Akka.Persistence.Linq2Db.BenchmarkTests/Docker/Linq2Db/DockerLinq2DbSqlServerJournalPerfSpec.cs
@@ -2,6 +2,9 @@
 using Akka.Configuration;
 using Akka.Persistence.Linq2Db.BenchmarkTests.Local.Linq2Db;
 using Akka.Persistence.Sql.Linq2Db;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Config;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types;
 using Akka.Persistence.Sql.Linq2Db.Tests;
 using LinqToDB;
 using Xunit;
