+++ src/Akka.Persistence.Linq2Db.BenchmarkTests/Local/Linq2Db/MSSQLiteLinq2DbJournalPerfSpec.cs
@@ -1,5 +1,8 @@
 ﻿using System;
 using Akka.Persistence.Sql.Linq2Db;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Config;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types;
 using Akka.Persistence.Sql.Linq2Db.Tests;
 using Akka.Util.Internal;
 using LinqToDB;
