+++ src/Akka.Persistence.Linq2Db.BenchmarkTests/Local/Linq2Db/SQLServerLinq2DbJournalPerfSpec.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using Akka.Configuration;
 using Akka.Persistence.Sql.Linq2Db;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Config;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types;
 using Akka.Persistence.Sql.Linq2Db.Tests;
 using JetBrains.dotMemoryUnit;
 using LinqToDB;
