+++ src/Akka.Persistence.Linq2Db.BenchmarkTests/Local/Linq2Db/SystemDataSQLiteLinq2DbJournalPerfSpec.cs
@@ -2,6 +2,9 @@
 using System.Data;
 using System.Data.SQLite;
 using Akka.Persistence.Sql.Linq2Db;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Config;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types;
 using Akka.Persistence.Sql.Linq2Db.Tests;
 using Akka.Util.Internal;
 using LinqToDB;
