+++ src/Akka.Persistence.Sql.Linq2Db.Journal/Journal/DAO/IJournalDaoWithUpdates.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Akka.Persistence.Sql.Linq2Db
+namespace Akka.Persistence.Sql.Linq2Db.Journal.Journal.DAO
 {
     public interface IJournalDaoWithUpdates : IJournalDao
     {
