+++ src/Akka.Persistence.Sql.Linq2Db.Journal/Journal/Types/JournalMetaData.cs
@@ -1,0 +1,12 @@
+﻿using LinqToDB.Mapping;
+
+namespace Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types
+{
+    public sealed class JournalMetaData
+    {
+        [Column(IsPrimaryKey = true, CanBeNull = false)]
+        public string PersistenceId { get; set; }
+        [PrimaryKey]
+        public long SequenceNumber { get; set; }
+    }
+}
