+++ src/Akka.Persistence.Sql.Linq2Db.Journal/Journal/Types/ReplayCompletion.cs
@@ -1,0 +1,8 @@
+﻿namespace Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types
+{
+    public class ReplayCompletion
+    {
+        public IPersistentRepresentation repr { get; set; }
+        public long SequenceNr { get; set; }
+    }
+}
