+++ src/Akka.Persistence.Sql.Linq2Db.Journal/Journal/Types/WriteFinished.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading.Tasks;
+
+namespace Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types
+{
+    public class WriteFinished
+    {
+        public WriteFinished(string persistenceId, Task future)
+        {
+            PersistenceId = persistenceId;
+            Future = future;
+        }
+        public string PersistenceId { get; protected set; }
+        public Task Future { get; protected set; }
+    }
+}
