+++ src/Akka.Persistence.Sql.Linq2Db.Journal/Journal/Types/WriteQueueSet.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using LanguageExt;
 
-namespace Akka.Persistence.Sql.Linq2Db
+namespace Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types
 {
     public class WriteQueueSet
     {
