+++ src/Akka.Persistence.Sql.Linq2Db.Tests/SQLServerJournalSpec.cs
@@ -1,5 +1,8 @@
 ﻿using System;
 using Akka.Configuration;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Config;
+using Akka.Persistence.Sql.Linq2Db.Journal.Journal.Types;
 using Akka.Persistence.TCK.Journal;
 using LinqToDB;
 using Xunit.Abstractions;
